Converts TimeGenerate to extension(Generate) block in Conjecture.Time
- Replaces TimeGenerate static class with TimeGenerateExtensions containing
  a C# 14 extension(Generate) block; users call Generate.TimeZones() and
  Generate.ClockSet() with a single using Conjecture.Time;
- Removes TimeGenerate entirely (pre-1.0 breaking change)
- Updates PublicAPI.Shipped/Unshipped, docs, README, and CHANGELOG
